Crate amethyst_core

A collection of structures and functions useful across the entire amethyst project.

Re-exports

pub use approx;

pub use nalgebra;

pub use shred;

pub use shrev;

pub use specs;

pub use crate::bundle::Error;

pub use crate::bundle::ErrorKind;

pub use crate::bundle::Result;

pub use crate::bundle::SystemBundle;

pub use crate::timing::*;

pub use crate::transform::*;

Modules

bundle
Provides a trait for adding bundles of systems to a dispatcher.

frame_limiter
Frame rate limiting.

timing
Utilities for working with time.

transform
amethyst transform ecs module

Structs

Named
A component that gives a name to an Entity.

Pausable
A system that is enabled when U has a specific value.

Enums

Axis2
A two dimensional axis.

Axis3
A three dimensional axis.

Traits

EventReader
Read events generically

SystemExt
Extension functionality associated systems.

WithNamed
An easy way to name an Entity and give it a Named Component.

Type Definitions

ArcThreadPool
A rayon thread pool wrapped in an Arc. This should be used as resource in World.