Type Definition amethyst_core::ecs::prelude::WriteExpect

type WriteExpect<'a, T> = Write<'a, T, PanicHandler>;

Allows to fetch a resource in a system mutably. This will panic if the resource does not exist. Usage of Write or Option<Write> is therefore recommended.