Enum amethyst_core::frame_limiter::FrameRateLimitStrategy

pub enum FrameRateLimitStrategy {
    Unlimited,
    Yield,
    Sleep,
    SleepAndYield(Duration),
}

Frame rate limiting strategy

Variants
Unlimited

No limit, will do a single yield, and then continue with the next frame

Yield

Use yield until the full frame duration has passed

Sleep

Use sleep until the full frame duration has passed

SleepAndYield(Duration)

Use sleep and yield combined, will use sleep strategy until the given duration remains, and then swap to yield strategy.