Struct amethyst_core::Float

pub struct Float(_);

A wrapper type around f32 and f64. It is used to hide the actual type being used internally. Mostly used with the Transform type. The default type is f32 and you can switch to the f64 type by enabling the "float64" feature gate.

Methods

impl Float

pub fn as_f32(self) -> f32

Get the internal value as a f32. Will cause a loss in precision if using the "float64" feature.

pub fn as_f64(self) -> f64

Get the internal value as a f64. Guaranteed to be lossless.