Struct amethyst_core::transform::components::LocalTransform

pub struct LocalTransform {
    pub rotation: Quaternion<f32>,
    pub scale: Vector3<f32>,
    pub translation: Vector3<f32>,
}

Local position, rotation, and scale (from parent if it exists).

Used for rendering position and orientation.

Fields
rotation: Quaternion<f32>

Quaternion [w (scalar), x, y, z]

scale: Vector3<f32>

Scale vector [x, y, z]

translation: Vector3<f32>

Translation/position vector [x, y, z]

Methods

impl LocalTransform

pub fn look_at(
&mut self,
orientation: &Orientation,
position: Point3<f32>
) -> &mut Self

Rotate to look at a point in space (without rolling)

pub fn matrix(&self) -> Matrix4<f32>

Returns the local object matrix for the transform.

Combined with the parent's global Transform component it gives the global (or world) matrix for the current entity.

pub fn move_forward(
&mut self,
orientation: &Orientation,
amount: f32
) -> &mut Self

Move relatively to its current position and orientation.

pub fn move_global(&mut self, direction: Vector3<f32>) -> &mut Self

Move relatively to its current position, but independently from its orientation. Ideally, first normalize the direction and then multiply it by whatever amount you want to move before passing the vector to this method

pub fn move_local(&mut self, axis: Vector3<f32>, amount: f32) -> &mut Self

Move relatively to its current position and orientation.

pub fn move_right(
&mut self,
orientation: &Orientation,
amount: f32
) -> &mut Self

Move relatively to its current position and orientation.

pub fn move_up(&mut self, orientation: &Orientation, amount: f32) -> &mut Self

Move relatively to its current position and orientation.

pub fn pitch_global(
&mut self,
orientation: &Orientation,
angle: Deg<f32>
) -> &mut Self

Pitch relatively to the world.

pub fn pitch_local(
&mut self,
orientation: &Orientation,
angle: Deg<f32>
) -> &mut Self

Pitch relatively to its own rotation.

pub fn roll_global(
&mut self,
orientation: &Orientation,
angle: Deg<f32>
) -> &mut Self

Roll relatively to the world.

pub fn roll_local(
&mut self,
orientation: &Orientation,
angle: Deg<f32>
) -> &mut Self

Roll relatively to its own rotation.

pub fn rotate(&mut self, quat: Quaternion<f32>) -> &mut Self

Add a rotation to the current rotation

pub fn rotate_global(
&mut self,
axis: Vector3<f32>,
angle: Deg<f32>
) -> &mut Self

Rotate relatively to the world

pub fn rotate_local(&mut self, axis: Vector3<f32>, angle: Deg<f32>) -> &mut Self

Rotate relatively to the current orientation

pub fn set_position(&mut self, position: Vector3<f32>) -> &mut Self

Set the position.

pub fn set_rotation<D: Into<Deg<f32>>>(&mut self, x: D, y: D, z: D) -> &mut Self

Set the rotation using Euler x, y, z.

pub fn to_view_matrix(&self, orientation: &Orientation) -> Matrix4<f32>

Calculate the view matrix from the given data.

pub fn yaw_global(
&mut self,
orientation: &Orientation,
angle: Deg<f32>
) -> &mut Self

Yaw relatively to the world.

pub fn yaw_local(
&mut self,
orientation: &Orientation,
angle: Deg<f32>
) -> &mut Self

Yaw relatively to its own rotation.

impl LocalTransform

pub fn new() -> Self

Create a new LocalTransform.

If you call matrix on this, then you would get an identity matrix.