Crate amethyst_core

A collection of structures and functions useful across the entire amethyst project.

Re-exports

pub use alga;

pub use approx;

pub use nalgebra as math;

pub use num_traits as num;

pub use specs as ecs;

pub use specs::shred;

pub use specs::shrev;

pub use crate::bundle::SystemBundle;

pub use crate::timing::*;

pub use crate::transform::*;

Modules

bundle
Provides a trait for adding bundles of systems to a dispatcher.

deferred_dispatcher_operation
Provides the ability to store Systems, Bundles, Barriers, in a normal vector for deferred dispatcher construction.

frame_limiter
Frame rate limiting.

geometry
Geometry helper functionality.

timing
Utilities for working with time.

transform
amethyst transform ecs module

Structs

Hidden
Hidden mesh component Useful for entities, that should not be rendered, but stay loaded in memory.

HiddenPropagate
Like Hidden, but can propagate through children when the HideHierarchySystem is enabled in the RenderBundle.

HideHierarchySystem
This system adds a HiddenPropagate-component to all children.

HideHierarchySystemDesc
Builds a HideHierarchySystem.

Named
A component that gives a name to an Entity.

Pausable
A system that is enabled when V has a specific value.

Enums

Axis2
A two dimensional axis.

Axis3
A three dimensional axis.

Traits

EventReader
Read events generically

RunNowDesc
Initializes a RunNow with some interaction with the World.

SystemDesc
Initializes a System with some interaction with the World.

SystemExt
Extension functionality associated systems.

WithNamed
An easy way to name an Entity and give it a Named Component.

Type Definitions

ArcThreadPool
A rayon thread pool wrapped in an Arc. This should be used as resource in World.