Struct amethyst_core::frame_limiter::FrameLimiter

pub struct FrameLimiter { /* fields omitted */ }

Frame limiter resource.

FrameLimiter is used internally by amethyst to limit the frame rate to the rate specified by the user. It is added as a resource to the world so that user code may change the frame rate limit at runtime if necessary.

Methods

impl FrameLimiter

pub fn new(strategy: FrameRateLimitStrategy, fps: u32) -> Self

Creates a new frame limiter.

pub fn set_rate(&mut self, strategy: FrameRateLimitStrategy, fps: u32)

Sets the maximum fps and frame rate limiting strategy.

pub fn from_config(config: FrameRateLimitConfig) -> Self

Creates a new frame limiter with the given config.

pub fn start(&mut self)

Resets the frame start time to the current instant.

This resets the frame limiter's internal tracking of when the last frame started to the current instant. Be careful when calling start, as doing so will cause the current frame to be longer than normal if not called at the very beginning of the frame.

pub fn wait(&mut self)

Blocks the current thread until the allotted frame time has passed.

wait is used internally by Application to limit the frame rate of the game to the configured rate. This should likely never be called directly by game logic.