Struct amethyst_core::transform::components::Transform

pub struct Transform(pub Matrix4<f32>);

Performs a global transformation on the entity (transform from origin).

Used for rendering position and orientation.

If this component is used, and TransformSystem is not used, then make sure to clear the flags on the FlaggedStorage at the appropriate times (before updating any Transform in the frame). See documentation on FlaggedStorage for more information.

Methods

impl Transform

pub fn is_finite(&self) -> bool

Checks whether each f32 of the Transform is finite (not NaN or inf).

impl Transform

pub fn new() -> Self

Creates a new Transform in the form of an identity matrix.